+++ server/src/lib/score-import/framework/goals/goals.ts
@@ -135,7 +135,7 @@ export async function ProcessGoal(
 					// that haven't changed return nothing instead of
 					// getting to this point.
 					lastInteraction: Date.now(),
-				},
+				} as Partial<GoalDocument>,
 			},
 		},
 	};